requirements.txt
Flask==2.2.5
Flask-Mail==0.9.1
Flask-Session==0.3.2
Flask-SocketIO==5.3.2
gunicorn==20.1.0
psycopg2-binary==2.9.7
fpdf==1.7.2
python-dotenv==0.21.0
requests==2.31.0
reportlab==3.6.12
wheel==0.40.0

# Google API and Authentication
gspread==5.2.0
oauth2client==4.1.3
google-auth==2.17.1
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
google-api-python-client==2.87.0

# Additional required packages
eventlet==0.33.3
apscheduler==3.9.1.post1
